// src/models/Book.ts
export interface Book {
    ID;
    Title: string;
    Description: string;
    PageCount: string;
    Excerpt: string;
    PublishDate;


    // "ID": 0,
    // "Title": "string",
    // "Description": "string",
    // "PageCount": 0,
    // "Excerpt": "string",
    // "PublishDate": '2018-07-31T18:49:56.712Z'
}